requirements.txt
rdkit-pypi>=2023.9.1
pandas>=2.0.0
numpy>=1.24.0
langgraph>=0.2.0
langchain-core>=0.3.0
langchain-anthropic>=0.3.0
